+++ Libraries/graphics/rectArray.cpp
@@ -1,24 +1,24 @@
 #include "graphics-internal.hpp"
 
-void gfx::RectArray::setRect(unsigned short index, RectShape rect) {
+void gfx::RectArray::setRect(int index, RectShape rect) {
     vertex_array[index * 4].position = sf::Vector2f(rect.x, rect.y);
     vertex_array[index * 4 + 1].position = sf::Vector2f(rect.x + rect.w, rect.y);
     vertex_array[index * 4 + 2].position = sf::Vector2f(rect.x + rect.w, rect.y + rect.h);
     vertex_array[index * 4 + 3].position = sf::Vector2f(rect.x, rect.y + rect.h);
 }
 
-void gfx::RectArray::setColor(unsigned short index, Color color) {
+void gfx::RectArray::setColor(int index, Color color) {
     vertex_array[index].color = {color.r, color.g, color.b, color.a};
 }
 
-void gfx::RectArray::setTextureCoords(unsigned short index, RectShape texture_coordinates) {
+void gfx::RectArray::setTextureCoords(int index, RectShape texture_coordinates) {
     vertex_array[index * 4].texCoords = sf::Vector2f(texture_coordinates.x, texture_coordinates.y);
     vertex_array[index * 4 + 1].texCoords = sf::Vector2f(texture_coordinates.x + texture_coordinates.w, texture_coordinates.y);
     vertex_array[index * 4 + 2].texCoords = sf::Vector2f(texture_coordinates.x + texture_coordinates.w, texture_coordinates.y + texture_coordinates.h);
     vertex_array[index * 4 + 3].texCoords = sf::Vector2f(texture_coordinates.x, texture_coordinates.y + texture_coordinates.h);
 }
 
-gfx::RectArray::RectArray(unsigned short size) {
+gfx::RectArray::RectArray(int size) {
     resize(size);
 }
